requirements.txt
anthropic==0.31.0
google-genai==0.6.0
httpx==0.27.0
python-dotenv==1.0.0
pydantic==2.7.1
